moved tests to correct location
